feat: improve cluster changes handling
Ensures batchers are stopped when a broker is removed from the cluster.
Batchers are automatically re-initialized when the broker re-joins the cluster.
This improves resilience and prevents errors due to unavailable brokers.

Also adds the possibility of defining if the client will work with unknown topics, improving the general use of the client.